requirements.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg[binary]>=3.2
django-filter>=24.2
python-dotenv>=1.0
